water-monitoring-iot/requirements.txt.txt
Django==3.2.25
djangorestframework==3.12.4
mysqlclient
pymysql